tools/requirements.txt
google-cloud-storage
json